deployment/backend/chatbot/requirements.txt
fastapi
uvicorn
google-generativeai
firebase-admin
pydantic
python-dotenv